+++ src/compiler/index.ts
@@ -59,9 +59,9 @@ export const constants = {
     PRESERVED_IDPREFIX
 }
 
-export { PositionFlag } from "./enums"
 export { isCompileError } from "./message/error"
 export { isCompileWarning } from "./message/warn"
+export { PositionFlag, TestingMode } from "./enums"
 export { parseDirectiveValue } from "./parser/directive"
 export { compile, compileIntermediate } from "./compile"
 export { parseEventFlagStandalone as parseEventFlag } from "./parser/event"