+++ src/compiler/transformer/runtime/codegen.ts
@@ -52,11 +52,6 @@ export function generateRuntimeCode(nodes: TemplateNode[]) {
     if (defaultProps) {
         writer.write(`${contextId}.P = `).writeScriptNode(defaultProps.value).wrapLine()
     }
-    if (inputDescriptor.styles.length) {
-        writer.write(
-            `${contextId}.o = ${getMaybeReusedString(` qk-${inputDescriptor.options.hashId}`)}\n`
-        )
-    }
     generateDelegateEventsRegistration(writer, contextId)
 
     if (usedIntrinsicVars.size) {